move: Dashboard Entity를 Dashboard 패키지로 이동(#32)
