+++ src/main/java/com/dnd/namuiwiki/domain/statistic/StatisticsService.java
@@ -2,7 +2,7 @@ package com.dnd.namuiwiki.domain.statistic;
 
 import com.dnd.namuiwiki.domain.dashboard.DashboardRepository;
 import com.dnd.namuiwiki.domain.statistic.model.Statistics;
-import com.dnd.namuiwiki.domain.statistic.model.entity.Dashboard;
+import com.dnd.namuiwiki.domain.dashboard.model.entity.Dashboard;
 import com.dnd.namuiwiki.domain.survey.model.entity.Survey;
 import com.dnd.namuiwiki.domain.survey.type.Period;
 import com.dnd.namuiwiki.domain.survey.type.Relation;
